Move TaskState model to workflows submodule
