+++ wagtail/models/workflows.py
@@ -7,6 +7,8 @@ from django.core.exceptions import PermissionDenied, ValidationError
 from django.db import models, transaction
 from django.db.models import Q
 from django.db.models.expressions import OuterRef, Subquery
+from django.utils import timezone
+from django.utils.functional import cached_property
 from django.utils.module_loading import import_string
 from django.utils.text import capfirst
 from django.utils.translation import gettext_lazy as _
@@ -20,6 +22,9 @@ from wagtail.forms import TaskStateCommentForm
 from wagtail.log_actions import log
 from wagtail.query import SpecificQuerySetMixin
 from wagtail.signals import (
+    task_approved,
+    task_cancelled,
+    task_rejected,
     task_submitted,
     workflow_approved,
     workflow_cancelled,
@@ -27,6 +32,7 @@ from wagtail.signals import (
     workflow_submitted,
 )
 
+from .copying import _copy, _copy_m2m_relations
 from .locking import LockableMixin
 from .orderable import Orderable
 from .revisions import Revision
@@ -254,8 +260,6 @@ class WorkflowState(models.Model):
         Checks the status of the current task, and progresses (or ends) the workflow if appropriate.
         If the workflow progresses, next_task will be used to start a specific task next if provided.
         """
-        from wagtail.models import TaskState
-
         if self.status != self.STATUS_IN_PROGRESS:
             # Updating a completed or cancelled workflow should have no effect
             return
@@ -293,8 +297,6 @@ class WorkflowState(models.Model):
 
     @property
     def successful_task_states(self):
-        from wagtail.models import TaskState
-
         successful_task_states = self.task_states.filter(
             Q(status=TaskState.STATUS_APPROVED) | Q(status=TaskState.STATUS_SKIPPED)
         )
@@ -318,7 +320,7 @@ class WorkflowState(models.Model):
 
     def cancel(self, user=None):
         """Cancels the workflow state"""
-        from wagtail.models import Page, TaskState
+        from wagtail.models import Page
 
         if self.status not in (self.STATUS_IN_PROGRESS, self.STATUS_NEEDS_CHANGES):
             raise PermissionDenied
@@ -369,8 +371,6 @@ class WorkflowState(models.Model):
         """
         Creates copies of previously approved task states with revision set to a different revision.
         """
-        from wagtail.models import TaskState
-
         approved_states = TaskState.objects.filter(
             workflow_state=self, status=TaskState.STATUS_APPROVED
         )
@@ -391,8 +391,6 @@ class WorkflowState(models.Model):
         """
         Returns the set of task states whose status applies to the current revision.
         """
-        from wagtail.models import TaskState
-
         task_states = TaskState.objects.filter(workflow_state_id=self.id)
         # If WAGTAIL_WORKFLOW_REQUIRE_REAPPROVAL_ON_EDIT=True, this is only task states created on the current revision
         if getattr(settings, "WAGTAIL_WORKFLOW_REQUIRE_REAPPROVAL_ON_EDIT", False):
@@ -415,8 +413,6 @@ class WorkflowState(models.Model):
         This is different to querying TaskState as it also returns tasks that haven't
         been started yet (so won't have a TaskState).
         """
-        from wagtail.models import TaskState
-
         # Get the set of task states whose status applies to the current revision
         task_states = self._get_applicable_task_states()
 
@@ -710,16 +706,12 @@ class Task(SpecificMixin, models.Model):
 
     @classmethod
     def get_task_state_class(self):
-        from wagtail.models import TaskState
-
         return self.task_state_class or TaskState
 
     def start(self, workflow_state, user=None):
         """
         Start this task on the provided workflow state by creating an instance of TaskState.
         """
-        from wagtail.models import TaskState
-
         task_state = self.get_task_state_class()(workflow_state=workflow_state)
         task_state.status = TaskState.STATUS_IN_PROGRESS
         task_state.revision = workflow_state.content_object.get_latest_revision()
@@ -789,8 +781,6 @@ class Task(SpecificMixin, models.Model):
 
     def get_task_states_user_can_moderate(self, user, **kwargs):
         """Returns a ``QuerySet`` of the task states the current user can moderate"""
-        from wagtail.models import TaskState
-
         return TaskState.objects.none()
 
     @classmethod
@@ -805,8 +795,6 @@ class Task(SpecificMixin, models.Model):
         """
         Set ``active`` to False and cancel all in progress task states linked to this task.
         """
-        from wagtail.models import TaskState
-
         self.active = False
         self.save()
         in_progress_states = TaskState.objects.filter(
@@ -892,8 +880,6 @@ class AbstractGroupApprovalTask(Task):
         return []
 
     def get_task_states_user_can_moderate(self, user, **kwargs):
-        from wagtail.models import TaskState
-
         if user.is_superuser or self._user_in_groups(user):
             return self.task_states.filter(status=TaskState.STATUS_IN_PROGRESS)
         else:
@@ -911,3 +897,256 @@ class AbstractGroupApprovalTask(Task):
 
 class GroupApprovalTask(AbstractGroupApprovalTask):
     pass
+
+
+class BaseTaskStateManager(models.Manager):
+    def reviewable_by(self, user):
+        tasks = Task.objects.filter(active=True).specific()
+        states = TaskState.objects.none()
+        for task in tasks:
+            states = states | task.get_task_states_user_can_moderate(user=user)
+        return states
+
+
+class TaskStateQuerySet(SpecificQuerySetMixin, models.QuerySet):
+    def for_instance(self, instance):
+        """
+        Filters to only TaskStates for the given instance
+        """
+        try:
+            # Use RevisionMixin.get_base_content_type() if available
+            return self.filter(
+                workflow_state__base_content_type=instance.get_base_content_type(),
+                workflow_state__object_id=str(instance.pk),
+            )
+        except AttributeError:
+            # Fallback to ContentType for the model
+            return self.filter(
+                workflow_state__content_type=ContentType.objects.get_for_model(
+                    instance, for_concrete_model=False
+                ),
+                workflow_state__object_id=str(instance.pk),
+            )
+
+
+TaskStateManager = BaseTaskStateManager.from_queryset(TaskStateQuerySet)
+
+
+class TaskState(SpecificMixin, models.Model):
+    """Tracks the status of a given Task for a particular revision."""
+
+    STATUS_IN_PROGRESS = "in_progress"
+    STATUS_APPROVED = "approved"
+    STATUS_REJECTED = "rejected"
+    STATUS_SKIPPED = "skipped"
+    STATUS_CANCELLED = "cancelled"
+    STATUS_CHOICES = (
+        (STATUS_IN_PROGRESS, _("In progress")),
+        (STATUS_APPROVED, _("Approved")),
+        (STATUS_REJECTED, _("Rejected")),
+        (STATUS_SKIPPED, _("Skipped")),
+        (STATUS_CANCELLED, _("Cancelled")),
+    )
+
+    workflow_state = models.ForeignKey(
+        "WorkflowState",
+        on_delete=models.CASCADE,
+        verbose_name=_("workflow state"),
+        related_name="task_states",
+    )
+    revision = models.ForeignKey(
+        "Revision",
+        on_delete=models.CASCADE,
+        verbose_name=_("revision"),
+        related_name="task_states",
+    )
+    task = models.ForeignKey(
+        "Task",
+        on_delete=models.CASCADE,
+        verbose_name=_("task"),
+        related_name="task_states",
+    )
+    status = models.fields.CharField(
+        choices=STATUS_CHOICES,
+        verbose_name=_("status"),
+        max_length=50,
+        default=STATUS_IN_PROGRESS,
+    )
+    started_at = models.DateTimeField(verbose_name=_("started at"), auto_now_add=True)
+    finished_at = models.DateTimeField(
+        verbose_name=_("finished at"), blank=True, null=True
+    )
+    finished_by = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        verbose_name=_("finished by"),
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name="finished_task_states",
+    )
+    comment = models.TextField(blank=True)
+    content_type = models.ForeignKey(
+        ContentType,
+        verbose_name=_("content type"),
+        related_name="wagtail_task_states",
+        on_delete=models.CASCADE,
+    )
+    exclude_fields_in_copy = []
+    default_exclude_fields_in_copy = ["id"]
+
+    objects = TaskStateManager()
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if not self.id:
+            # this model is being newly created
+            # rather than retrieved from the db;
+            if not self.content_type_id:
+                # set content type to correctly represent the model class
+                # that this was created as
+                self.content_type = ContentType.objects.get_for_model(self)
+
+    def __str__(self):
+        return _("Task '%(task_name)s' on Revision '%(revision_info)s': %(status)s") % {
+            "task_name": self.task,
+            "revision_info": self.revision,
+            "status": self.status,
+        }
+
+    @transaction.atomic
+    def approve(self, user=None, update=True, comment=""):
+        """
+        Approve the task state and update the workflow state.
+        """
+        if self.status != self.STATUS_IN_PROGRESS:
+            raise PermissionDenied
+        self.status = self.STATUS_APPROVED
+        self.finished_at = timezone.now()
+        self.finished_by = user
+        self.comment = comment
+        self.save()
+
+        self.log_state_change_action(user, "approve")
+        if update:
+            self.workflow_state.update(user=user)
+        task_approved.send(
+            sender=self.specific.__class__, instance=self.specific, user=user
+        )
+        return self
+
+    @transaction.atomic
+    def reject(self, user=None, update=True, comment=""):
+        """
+        Reject the task state and update the workflow state.
+        """
+        if self.status != self.STATUS_IN_PROGRESS:
+            raise PermissionDenied
+        self.status = self.STATUS_REJECTED
+        self.finished_at = timezone.now()
+        self.finished_by = user
+        self.comment = comment
+        self.save()
+
+        self.log_state_change_action(user, "reject")
+        if update:
+            self.workflow_state.update(user=user)
+        task_rejected.send(
+            sender=self.specific.__class__, instance=self.specific, user=user
+        )
+
+        return self
+
+    @cached_property
+    def task_type_started_at(self):
+        """
+        Finds the first chronological started_at for successive TaskStates - ie started_at if the task had not been restarted.
+        """
+        task_states = (
+            TaskState.objects.filter(workflow_state=self.workflow_state)
+            .order_by("-started_at")
+            .select_related("task")
+        )
+        started_at = None
+        for task_state in task_states:
+            if task_state.task == self.task:
+                started_at = task_state.started_at
+            elif started_at:
+                break
+        return started_at
+
+    @transaction.atomic
+    def cancel(self, user=None, resume=False, comment=""):
+        """
+        Cancel the task state and update the workflow state.
+        If ``resume`` is set to True, then upon update the workflow state is passed the current task as ``next_task``,
+        causing it to start a new task state on the current task if possible.
+        """
+        self.status = self.STATUS_CANCELLED
+        self.finished_at = timezone.now()
+        self.comment = comment
+        self.finished_by = user
+        self.save()
+        if resume:
+            self.workflow_state.update(user=user, next_task=self.task.specific)
+        else:
+            self.workflow_state.update(user=user)
+        task_cancelled.send(
+            sender=self.specific.__class__, instance=self.specific, user=user
+        )
+        return self
+
+    def copy(self, update_attrs=None, exclude_fields=None):
+        """
+        Copy this task state, excluding the attributes in the ``exclude_fields`` list and updating any attributes
+        to values specified in the ``update_attrs`` dictionary of ``attribute``: ``new value`` pairs.
+        """
+        exclude_fields = (
+            self.default_exclude_fields_in_copy
+            + self.exclude_fields_in_copy
+            + (exclude_fields or [])
+        )
+        instance, child_object_map = _copy(self.specific, exclude_fields, update_attrs)
+        instance.save()
+        _copy_m2m_relations(self, instance, exclude_fields=exclude_fields)
+        return instance
+
+    def get_comment(self):
+        """
+        Returns a string that is displayed in workflow history.
+
+        This could be a comment by the reviewer, or generated.
+        Use mark_safe to return HTML.
+        """
+        return self.comment
+
+    def log_state_change_action(self, user, action):
+        """Log the approval/rejection action"""
+        obj = self.revision.as_object()
+        next_task = self.workflow_state.get_next_task()
+        next_task_data = None
+        if next_task:
+            next_task_data = {"id": next_task.id, "title": next_task.name}
+        log(
+            instance=obj,
+            action=f"wagtail.workflow.{action}",
+            user=user,
+            data={
+                "workflow": {
+                    "id": self.workflow_state.workflow.id,
+                    "title": self.workflow_state.workflow.name,
+                    "status": self.status,
+                    "task_state_id": self.id,
+                    "task": {
+                        "id": self.task.id,
+                        "title": self.task.name,
+                    },
+                    "next": next_task_data,
+                },
+                "comment": self.get_comment(),
+            },
+            revision=self.revision,
+        )
+
+    class Meta:
+        verbose_name = _("Task state")
+        verbose_name_plural = _("Task states")
